Add files via upload
Created a password manager that lets users store, access and organize their usernames and passwords without using their hands. It uses Python for voice recognition so users can interact with it using voice commands. SQL is used to securely and encrypt store the credentials. The main goal of this project is to make password management more convenient and secure by allowing users to manage their passwords through voice commands for a smoother experience.